+++ src/main/java/edu/kit/datamanager/indexer/messagehandler/MetastoreMessageHandler.java
@@ -24,12 +24,17 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import com.github.jknack.handlebars.Handlebars;
+import com.google.gson.Gson;
+import com.google.gson.JsonObject;
+import edu.kit.datamanager.clients.SimpleServiceClient;
 import edu.kit.datamanager.indexer.configuration.ApplicationProperties;
 import edu.kit.datamanager.indexer.exception.IndexerException;
 import edu.kit.datamanager.indexer.service.impl.IndexingService;
 import edu.kit.datamanager.indexer.service.impl.MappingService;
 import edu.kit.datamanager.indexer.util.ElasticsearchUtil;
+import java.io.ByteArrayOutputStream;
 import java.io.IOException;
+import java.io.OutputStream;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.nio.charset.StandardCharsets;
@@ -47,6 +52,8 @@ public class MetastoreMessageHandler implements IMessageHandler {
 
   private static final Logger LOG = LoggerFactory.getLogger(MetastoreMessageHandler.class);
 
+  public final static String RESOURCE_TYPE = "application/vnd.datamanager.metadata-record+json";
+
   private Handlebars hb;
 
   @Autowired
@@ -60,9 +67,10 @@ public class MetastoreMessageHandler implements IMessageHandler {
 
   @Autowired
   private IConsumerEngine consumer;
-  
+
   public static final String RESOURCE_URL = "resolvingUrl";
   public static final String MAPPING_ID = "documentType";
+  private static final String MAPPING_URL_APPENDIX = "/api/v1/schemas/";
 
   MetastoreMessageHandler(ApplicationProperties configuration, MappingService mappingService, IndexingService indexingService) {
     properties = configuration;
@@ -76,6 +84,7 @@ public class MetastoreMessageHandler implements IMessageHandler {
 
     String resourceUrlAsString = message.getMetadata().get(RESOURCE_URL);
     String mappingId = message.getMetadata().get(MAPPING_ID);
+
     if ((resourceUrlAsString == null) || (mappingId == null)) {
       LOG.debug("Reject message: Missing properties!");
       return RESULT.REJECTED;
@@ -84,6 +93,24 @@ public class MetastoreMessageHandler implements IMessageHandler {
       LOG.debug("Reject message: Not addressed correctly!");
       return RESULT.REJECTED;
     }
+    String schemaId = mappingId;
+    int beginIndex = mappingId.indexOf(MAPPING_URL_APPENDIX);
+    int endIndex = mappingId.indexOf("?", beginIndex);
+    if (beginIndex > 0) {
+      beginIndex += MAPPING_URL_APPENDIX.length();
+    } else {
+      beginIndex = 0;
+    }
+    if (endIndex < 0) {
+      endIndex = mappingId.length();
+    }
+    schemaId = mappingId.substring(beginIndex, endIndex);
+    LOG.trace("Strip mappingId to id only...");
+    mappingId = mappingId.substring(beginIndex);
+    mappingId = mappingId.replaceAll("=", "_");
+    LOG.trace("New mappingId: '{}'", mappingId);
+    LOG.trace("SchemaId: '{}'", schemaId);
+
     Path resultPath = null;
     try {
       LOG.debug("This message is for me: {}", message);
@@ -95,10 +122,12 @@ public class MetastoreMessageHandler implements IMessageHandler {
         return RESULT.FAILED;
       }
       resultPath = pathWithAllMappings.get(0);
-      String index = ElasticsearchUtil.testForValidIndex(mappingId);
-      if (index.equals(mappingId)) {
-        LOG.warn("MappingId '{}' was transformed to '{}' due to restrictions of elasticsearch!", mappingId, index);
-      }
+      // Prefix index with configured value. (default: metastore-
+      String index = properties.getElasticsearchIndex() + schemaId;
+//      String index = ElasticsearchUtil.testForValidIndex(mappingId);
+//      if (!index.equals(mappingId)) {
+//        LOG.warn("MappingId '{}' was transformed to '{}' due to restrictions of elasticsearch!", mappingId, index);
+//      }
 
       String jsonDocument = FileUtils.readFileToString(resultPath.toFile(), StandardCharsets.UTF_8);
       elasticsearchService.uploadToElastic(jsonDocument, index, properties.getElasticsearchType(), resourceUrlAsString);